Restruture files
